Implement multi-value charmaps

+++ include/util.hpp
@@ -12,6 +12,6 @@ char const *printChar(int c);
 /*
  * @return The number of bytes read, or 0 if invalid data was found
  */
-size_t readUTF8Char(std::vector<uint8_t> *dest, char const *src);
+size_t readUTF8Char(std::vector<int32_t> *dest, char const *src);
 
 #endif // RGBDS_UTIL_HPP
